move id generator to core
