+++ src/vs/base/common/idGenerator.ts
